requirements.txt
africastalking==1.2.3
amqp==5.0.1
apipkg==1.5
asgiref==3.2.10
attrs==20.2.0
autopep8==1.5.5
billiard==3.6.3.0
celery==5.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
coverage-badge==1.0.1
cryptography==3.1.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.1
django-extensions==2.2.5
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.1
drf-spectacular==0.10.0
drf-yasg==1.20.0
execnet==1.8.0
Faker==6.5.0
flake8==3.8.3
gunicorn==20.0.4
hypothesis==6.4.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
iniconfig==1.0.1
install==1.3.3
itypes==1.2.0
Jinja2==2.11.2
josepy==1.4.0
jsonschema==3.2.0
kombu==5.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
mozilla-django-oidc==1.2.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pep8==1.7.1
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.1.0
pytest-cov==2.10.1
pytest-django==3.10.0
pytest-forked==1.3.0
pytest-xdist==2.2.1
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.7.2
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sortedcontainers==2.3.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
wcwidth==0.2.5
Werkzeug==0.16.0
whitenoise==5.2.0
zipp==3.2.0
